// src/api/invoices/api.ts
export interface GenerateInvoiceRequest {
    /**
     * 
     * @type {string}
     * @memberof GenerateInvoiceRequest
     */
    'invoiceNumber': string;
    /**
     * 
     * @type {string}
     * @memberof GenerateInvoiceRequest
     */
    'invoiceDueDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof GenerateInvoiceRequest
     */
    'clientId': string;
    /**
     * 
     * @type {Array<ProductRequest>}
     * @memberof GenerateInvoiceRequest
     */
    'products': Array<ProductRequest>;
    /**
     * 
     * @type {string}
     * @memberof GenerateInvoiceRequest
     */
    'paymentTerms'?: string;
    /**
     * 
     * @type {string}
     * @memberof GenerateInvoiceRequest
     */
    'observations'?: string;
    /**
     * 
     * @type {string}
     * @memberof GenerateInvoiceRequest
     */
    'templateId'?: string;
    /**
     * 
     * @type {string}
     * @memberof GenerateInvoiceRequest
     */
    'kind'?: GenerateInvoiceRequestKindEnum;
    /**
     * 
     * @type {number}
     * @memberof GenerateInvoiceRequest
     */
    'discount'?: number;
    /**
     * 
     * @type {number}
     * @memberof GenerateInvoiceRequest
     */
    'tax'?: number;
}

export const GenerateInvoiceRequestKindEnum = {
    F: 'F',
    E: 'E'
} as const;

export type GenerateInvoiceRequestKindEnum = typeof GenerateInvoiceRequestKindEnum[keyof typeof GenerateInvoiceRequestKindEnum];

/**
 * 
 * @export
 * @interface GeneratedInvoiceResponse
 */
export interface GeneratedInvoiceResponse {
    /**
     * 
     * @type {string}
     * @memberof GeneratedInvoiceResponse
     */
    'documentUrl'?: string;
    /**
     * 
     * @type {string}
     * @memberof GeneratedInvoiceResponse
     */
    'publicUrl'?: string;
}
/**
 * 
 * @export
 * @interface GetAllInvoicesPaginatedResponse
 */
export interface GetAllInvoicesPaginatedResponse {
    /**
     * 
     * @type {Pagination}
     * @memberof GetAllInvoicesPaginatedResponse
     */
    'pagination'?: Pagination;
    /**
     * 
     * @type {Array<InvoiceResponse>}
     * @memberof GetAllInvoicesPaginatedResponse
     */
    'data'?: Array<InvoiceResponse>;
}
// ...
export interface Pagination {
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'totalItems'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'pageSize'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'currentPage'?: number;
    /**
     * 
     * @type {number}
     * @memberof Pagination
     */
    'totalPages'?: number;
}
/**
 * 
 * @export
 * @interface ProductRequest
 */
export interface ProductRequest {
    /**
     * 
     * @type {string}
     * @memberof ProductRequest
     */
    'productId': string;
    /**
     * 
     * @type {number}
     * @memberof ProductRequest
     */
    'price': number;
    /**
     * 
     * @type {number}
     * @memberof ProductRequest
     */
    'quantity': number;
    /**
     * 
     * @type {Array<string>}
     * @memberof ProductRequest
     */
    'fees'?: Array<string>;
}
/**
 * 
 * @export
 * @interface SendInvoiceByEmail200Response
 */
export interface SendInvoiceByEmail200Response {
    /**
     * 
     * @type {string}
     * @memberof SendInvoiceByEmail200Response
     */
    'message'?: string;
}
/**
 * 
 * @export
 * @interface SendInvoiceEmailRequest
 */
export interface SendInvoiceEmailRequest {
    /**
     * The unique identifier of the client.
     * @type {string}
     * @memberof SendInvoiceEmailRequest
     */
    'clientId': string;
    /**
     * The unique identifier of the invoice to send.
     * @type {string}
     * @memberof SendInvoiceEmailRequest
     */
    'invoiceId': string;
    /**
     * Optional message to include in the email.
     * @type {string}
     * @memberof SendInvoiceEmailRequest
     */
    'message': string;
}
// ...
export interface InvoiceApiDeleteInvoiceRequest {
    /**
     * The unique identifier of the invoice to delete.
     * @type {string}
     * @memberof InvoiceApiDeleteInvoice
     */
    readonly invoiceId: string
}

/**
 * Request parameters for downloadInvoice operation in InvoiceApi.
 * @export
 * @interface InvoiceApiDownloadInvoiceRequest
 */
export interface InvoiceApiDownloadInvoiceRequest {
    /**
     * The unique identifier of the invoice.
     * @type {string}
     * @memberof InvoiceApiDownloadInvoice
     */
    readonly invoiceId: string
}

/**
 * Request parameters for generateInvoice operation in InvoiceApi.
 * @export
 * @interface InvoiceApiGenerateInvoiceRequest
 */
export interface InvoiceApiGenerateInvoiceRequest {
    /**
     * 
     * @type {GenerateInvoiceRequest}
     * @memberof InvoiceApiGenerateInvoice
     */
    readonly generateInvoiceRequest: GenerateInvoiceRequest
}

/**
 * Request parameters for generateInvoiceReference operation in InvoiceApi.
 * @export
 * @interface InvoiceApiGenerateInvoiceReferenceRequest
 */
export interface InvoiceApiGenerateInvoiceReferenceRequest {
    /**
     * 
     * @type {'F' | 'E'}
     * @memberof InvoiceApiGenerateInvoiceReference
     */
    readonly kind?: GenerateInvoiceReferenceKindEnum
}

/**
 * Request parameters for getAllInvoices operation in InvoiceApi.
 * @export
 * @interface InvoiceApiGetAllInvoicesRequest
 */
export interface InvoiceApiGetAllInvoicesRequest {
    /**
     * The starting index of the invoices.
     * @type {number}
     * @memberof InvoiceApiGetAllInvoices
     */
    readonly from?: number

    /**
     * The number of invoices to return.
     * @type {number}
     * @memberof InvoiceApiGetAllInvoices
     */
    readonly size?: number

    /**
     * Filter criteria for invoices.
     * @type {string}
     * @memberof InvoiceApiGetAllInvoices
     */
    readonly filter?: string

    /**
     * Field to order invoices by.
     * @type {'createdAt' | 'views'}
     * @memberof InvoiceApiGetAllInvoices
     */
    readonly orderBy?: GetAllInvoicesOrderByEnum

    /**
     * Sorting direction.
     * @type {'ASC' | 'DESC'}
     * @memberof InvoiceApiGetAllInvoices
     */
    readonly direction?: GetAllInvoicesDirectionEnum

    /**
     * Invoice kind Final (F) or Estimation (E).
     * @type {'F' | 'E'}
     * @memberof InvoiceApiGetAllInvoices
     */
    readonly kind?: GetAllInvoicesKindEnum

    /**
     * Filter by settlement status of the invoice.
     * @type {boolean}
     * @memberof InvoiceApiGetAllInvoices
     */
    readonly settled?: boolean

    /**
     * Filter invoices by their status.
     * @type {'ACTIVE' | 'EXPIRED' | 'PAID' | 'UNPAID' | 'INACTIVE'}
     * @memberof InvoiceApiGetAllInvoices
     */
    readonly status?: GetAllInvoicesStatusEnum
}

/**
 * Request parameters for sendInvoiceByEmail operation in InvoiceApi.
 * @export
 * @interface InvoiceApiSendInvoiceByEmailRequest
 */
export interface InvoiceApiSendInvoiceByEmailRequest {
    /**
     * 
     * @type {SendInvoiceEmailRequest}
     * @memberof InvoiceApiSendInvoiceByEmail
     */
    readonly sendInvoiceEmailRequest: SendInvoiceEmailRequest
}

/**
 * InvoiceApi - object-oriented interface
 * @export
 * @class InvoiceApi
 * @extends {BaseAPI}
 */
export class InvoiceApi extends BaseAPI {
    /**
     * Deletes a specific invoice by its unique identifier. Requires authentication.
     * @summary Delete an invoice
     * @param {InvoiceApiDeleteInvoiceRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof InvoiceApi
     */
    public deleteInvoice(requestParameters: InvoiceApiDeleteInvoiceRequest, options?: RawAxiosRequestConfig) {
        return InvoiceApiFp(this.configuration).deleteInvoice(requestParameters.invoiceId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Retrieve a specific invoice in PDF format.
     * @summary Download invoice as PDF
     * @param {InvoiceApiDownloadInvoiceRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof InvoiceApi
     */
    public downloadInvoice(requestParameters: InvoiceApiDownloadInvoiceRequest, options?: RawAxiosRequestConfig) {
        return InvoiceApiFp(this.configuration).downloadInvoice(requestParameters.invoiceId, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary Generate an invoice and send its document
     * @param {InvoiceApiGenerateInvoiceRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof InvoiceApi
     */
    public generateInvoice(requestParameters: InvoiceApiGenerateInvoiceRequest, options?: RawAxiosRequestConfig) {
        return InvoiceApiFp(this.configuration).generateInvoice(requestParameters.generateInvoiceRequest, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Returns a unique reference string to be used in invoice creation.
     * @summary Generate an invoice reference
     * @param {InvoiceApiGenerateInvoiceReferenceRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof InvoiceApi
     */
    public generateInvoiceReference(requestParameters: InvoiceApiGenerateInvoiceReferenceRequest = {}, options?: RawAxiosRequestConfig) {
        return InvoiceApiFp(this.configuration).generateInvoiceReference(requestParameters.kind, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * Retrieve a list of invoices with optional filters, sorting, and pagination. Requests require JWT and Refresh Tokens to be present in cookies.
     * @summary Fetch paginated invoices
     * @param {InvoiceApiGetAllInvoicesRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof InvoiceApi
     */
    public getAllInvoices(requestParameters: InvoiceApiGetAllInvoicesRequest = {}, options?: RawAxiosRequestConfig) {
       return InvoiceApiFp(this.configuration)
    .getAllInvoices(
        requestParameters.from,
        requestParameters.size,
        requestParameters.filter,
        requestParameters.orderBy,
        requestParameters.direction,
        requestParameters.kind,
        requestParameters.settled,
        requestParameters.status,
        options
    )
    .then((request) => request(this.axios, this.basePath));

    }

    /**
     * Sends an existing invoice to the specified client\'s email address with an optional message.
     * @summary Send invoice via email
     * @param {InvoiceApiSendInvoiceByEmailRequest} requestParameters Request parameters.
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof InvoiceApi
     */
    public sendInvoiceByEmail(requestParameters: InvoiceApiSendInvoiceByEmailRequest, options?: RawAxiosRequestConfig) {
        return InvoiceApiFp(this.configuration).sendInvoiceByEmail(requestParameters.sendInvoiceEmailRequest, options).then((request) => request(this.axios, this.basePath));
    }
}

/**
 * @export
 */
export const GenerateInvoiceReferenceKindEnum = {
    F: 'F',
    E: 'E'
} as const;
export type GenerateInvoiceReferenceKindEnum = typeof GenerateInvoiceReferenceKindEnum[keyof typeof GenerateInvoiceReferenceKindEnum];
/**
 * @export
 */
export const GetAllInvoicesOrderByEnum = {
    CreatedAt: 'createdAt',
    Views: 'views'
} as const;
export type GetAllInvoicesOrderByEnum = typeof GetAllInvoicesOrderByEnum[keyof typeof GetAllInvoicesOrderByEnum];
/**
 * @export
 */
export const GetAllInvoicesDirectionEnum = {
    Asc: 'ASC',
    Desc: 'DESC'
} as const;
export type GetAllInvoicesDirectionEnum = typeof GetAllInvoicesDirectionEnum[keyof typeof GetAllInvoicesDirectionEnum];
/**
 * @export
 */
export const GetAllInvoicesKindEnum = {
    F: 'F',
    E: 'E'
} as const;
export type GetAllInvoicesKindEnum = typeof GetAllInvoicesKindEnum[keyof typeof GetAllInvoicesKindEnum];
/**
 * @export
 */
export const GetAllInvoicesStatusEnum = {
    Active: 'ACTIVE',
    Expired: 'EXPIRED',
    Paid: 'PAID',
    Unpaid: 'UNPAID',
    Inactive: 'INACTIVE'
} as const;
export type GetAllInvoicesStatusEnum = typeof GetAllInvoicesStatusEnum[keyof typeof GetAllInvoicesStatusEnum];